add specialist references
