+++ blueprints/__init__.py
@@ -1,4 +1,6 @@
 
 from .uk_who_blueprint import *
+from .trisomy_21_blueprint import *
+from .turner_blueprint import *
 from .utilities_blueprint import *
 from .openapi_blueprint import *
