--- controllers/fictional_children.py
@@ -1,11 +1,0 @@
-import rcpchgrowth.rcpchgrowth as rcpchgrowth
-
-def generate_fictional_children_data(
-        measurement_method,
-        sex
-    ):
-
-    return rcpchgrowth.generate_fictional_children_data(
-        measurement_method=measurement_method,
-        sex=sex
-    )
