+++ openapi.yml
@@ -88,68 +88,6 @@ components:
         observation_value:
           type: number
       type: object
-    FictionalChildRequestParameters:
-      properties:
-        drift_amount:
-          description: The SDS range over which you want serial generated values to
-            drift, simulating growth or pathological state
-          type: string
-        interval_type:
-          description: The length of the time intervals used. Can be `d`, `day`, `days`,
-            `weeks`, `m`, `month`, `months`, `y`, `year`, `years`. For example, setting
-            the time interval to `months` and the intervals to `3` would generate
-            fictional child growth data at 3-monthly intervals
-          enum:
-          - d
-          - day
-          - days
-          - weeks
-          - m
-          - month
-          - months
-          - y
-          - year
-          - years
-          type: string
-        intervals:
-          description: The value of the time intervals between generated measurements.
-            Set the time intervals using the `interval_type` parameter.
-          format: int32
-          type: integer
-        measurement_method:
-          description: 'The type of measurement requested: can be `height`, `weight`,
-            `ofc` or `bmi`'
-          enum:
-          - height
-          - weight
-          - ofc
-          - bmi
-          type: string
-        number_of_measurements:
-          description: The number of sequential measurements you would like the API
-            to generate, at intervals specified by the `intervals` and `interval_type`
-            parameters.
-          format: int32
-          type: integer
-        sex:
-          description: The sex of the child
-          enum:
-          - male
-          - female
-          type: string
-        starting_age:
-          description: The **decimal** age at which the fictional child data series
-            should start. Decimal ages are the child's age in years as a floating
-            point number
-          format: float
-          type: number
-        starting_sds:
-          description: The starting SDS at which you want the API to generate fictional
-            measurements. This SDS values will drift upwards or downwards over the
-            time series, according to the `drift_amount` parameter.
-          format: float
-          type: number
-      type: object
     MeasurementCalculatedValues:
       properties:
         centile:
@@ -228,23 +166,11 @@ components:
           description: URL for the publication (if available)
           type: string
       type: object
-    calculation:
-      properties:
-        calculation:
-          $ref: '#/components/schemas/MeasurementResponse'
-      type: object
     chartData:
       properties:
         centile_data:
           type: string
-        child_data:
-          type: string
         sex:
-          type: string
-      type: object
-    fictionalChild:
-      properties:
-        response:
           type: string
       type: object
     plottableChildData:
@@ -260,6 +186,16 @@ components:
           items:
             $ref: '#/components/schemas/Reference'
           type: array
+      type: object
+    trisomy_21_calculation:
+      properties:
+        calculation:
+          $ref: '#/components/schemas/MeasurementResponse'
+      type: object
+    uk_who_calculation:
+      properties:
+        calculation:
+          $ref: '#/components/schemas/MeasurementResponse'
       type: object
 info:
   description: Royal College of Paediatrics and Child Health Digital Growth Charts
@@ -301,21 +237,12 @@ paths:
           content:
             application/json:
               schema:
-                $ref: '#/components/schemas/calculation'
+                $ref: '#/components/schemas/uk_who_calculation'
           description: Centile calculation (single) according to the supplied data
             was returned
       summary: Centile and SDS Calculation route.
   /uk-who/chart-coordinates:
     post:
-      description: '* Requires calculated child measurement results from a preceding
-        call to the `uk-who/calculation` endpoint.
-
-        * Pass these results back in as the JSON body payload.
-
-        * Returns coordinates for constructing the lines of a traditional growth chart,
-        in JSON format
-
-        '
       requestBody:
         content:
           application/json:
@@ -328,8 +255,8 @@ paths:
               schema:
                 $ref: '#/components/schemas/chartData'
           description: Chart data for plotting a traditional growth chart was returned
-      summary: UK-WHO Chart coordinates in plottable format, including the plottable
-        child measuremnt coordinates
+      summary: UK-WHO Chart coordinates in plottable format * Returns coordinates
+        for constructing the lines of a traditional growth chart, in JSON format
   /uk-who/plottable-child-data:
     post:
       description: '* Requires results data parameters from a call to the calculation
@@ -374,27 +301,6 @@ paths:
                 $ref: '#/components/schemas/references'
           description: Reference data was returned
       summary: Centile References Library API route.
-  /utilities/create_fictional_child_measurements:
-    post:
-      description: '* Returns a series of generated fictional measurement data for
-        a child.
-
-        * Used for testing, demonstration and research purposes.
-
-        '
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/FictionalChildRequestParameters'
-      responses:
-        '200':
-          content:
-            application/json:
-              schema:
-                $ref: '#/components/schemas/fictionalChild'
-          description: Fictional child test data was returned
-      summary: Fictional Child Data Generator API route.
   /utilities/instructions:
     get:
       description: '* Does not expect any parameters.
@@ -412,6 +318,41 @@ paths:
                 type: string
           description: API Instructions and information was returned
       summary: Instructions API route.
+  /trisomy-21/calculation:
+    post:
+      description: '* Returns a single centile/SDS calculation for the selected `measurement_method`.
+
+        * Gestational age correction will be applied automatically if appropriate
+        according to the gestational age at birth data supplied.
+
+        * Available `measurement_method`s are: `height`, `weight`, `bmi`, or `ofc`
+        (OFC = occipitofrontal circumference = ''head circumference'').
+
+        * Note that BMI must be precalculated for the `bmi` function.
+
+        '
+      requestBody:
+        content:
+          application/json:
+            example:
+              birth_date: '2020-04-12'
+              gestation_days: 4
+              gestation_weeks: 40
+              measurement_method: height
+              observation_date: '2020-06-12'
+              observation_value: 60
+              sex: male
+            schema:
+              $ref: '#/components/schemas/CalculationRequestParameters'
+      responses:
+        '200':
+          content:
+            application/json:
+              schema:
+                $ref: '#/components/schemas/trisomy_21_calculation'
+          description: Centile calculation (single) according to the supplied data
+            was returned
+      summary: Centile and SDS Calculation route.
   /:
     get:
       description: '* The root endpoint of the Digital Growth Charts API returns the
