+++ rcpchgrowth/rcpchgrowth/__init__.py
@@ -1,9 +1,11 @@
 from .date_calculations import decimal_age, chronological_decimal_age, corrected_decimal_age, chronological_calendar_age, estimated_date_delivery, corrected_gestational_age
-from .global_functions import centile, sds_for_measurement, measurement_from_sds, percentage_median_bmi, create_uk_who_chart, create_trisomy_21_chart, create_turner_chart
+from .global_functions import centile, sds_for_measurement, measurement_from_sds, percentage_median_bmi
 from .centile_bands import centile_band_for_centile
 from .bmi_functions import bmi_from_height_weight, weight_for_bmi_height
 from .growth_interpretations import comment_prematurity_correction
-from .dynamic_growth import velocity, acceleration, correlate_weight, create_fictional_child
+from .uk_who import select_reference_data_for_uk_who_chart
+from .turner import select_reference_data_for_turner
+from .trisomy_21 import select_reference_data_for_trisomy_21
 from .measurement import Measurement
-from .fictional_children import generate_fictional_children_data
+from .chart_functions import create_chart, create_plottable_child_data
 from .constants import *
