+++ rcpchgrowth/rcpchgrowth/global_functions.py
@@ -1,12 +1,12 @@
 import math
 import scipy.stats as stats
 from scipy.interpolate import interp1d
+from .uk_who import uk_who_lms_array_for_measurement_and_sex
+from .turner import turner_lms_array_for_measurement_and_sex
+from .trisomy_21 import trisomy_21_lms_array_for_measurement_and_sex
 # from scipy import interpolate  #see below, comment back in if swapping interpolation method
 # from scipy.interpolate import CubicSpline #see below, comment back in if swapping interpolation method
-from .uk_who import uk_who_lms_array_for_measurement_and_sex, select_reference_data_for_uk_who_chart
-from .turner import turner_lms_array_for_measurement_and_sex, select_reference_data_for_turners
-from .trisomy_21 import trisomy_21_lms_array_for_measurement_and_sex, select_reference_data_for_trisomy_21
-from .constants.parameter_constants import *
+from .constants.parameter_constants import UK_WHO, TURNERS, TRISOMY_21, COLE_TWO_THIRDS_SDS_NINE_CENTILES, COLE_TWO_THIRDS_SDS_NINE_CENTILE_COLLECTION, THREE_PERCENT_CENTILE_COLLECTION, MEASUREMENT_METHODS, SEXES, UK_WHO_REFERENCES
 import logging
 import json
 
@@ -269,33 +269,6 @@ def percentage_median_bmi(reference: str, age: float, actual_bmi: float, sex: st
     percent_median_bmi = (actual_bmi / m) * 100.0
     return percent_median_bmi
 
-
-def lms_value_array_for_measurement_for_reference(
-    reference: str,
-    age: float,
-    measurement_method: str,
-    sex: str,
-    born_preterm: bool
-) -> list:
-    """
-    This is a private function which returns the LMS array for measurement_method and sex and reference
-    It accepts the reference ('uk-who', 'turners-syndrome' or 'trisomy-21')
-    """
-
-    if reference == "uk-who":
-        lms_value_array_for_measurement = uk_who_lms_array_for_measurement_and_sex(
-            age=age, measurement_method=measurement_method, sex=sex, born_preterm=born_preterm)
-    elif reference == "turners-syndrome":
-        lms_value_array_for_measurement = turner_lms_array_for_measurement_and_sex(
-            measurement_method=measurement_method, sex=sex, age=age)
-    elif reference == "trisomy-21":
-        lms_value_array_for_measurement = trisomy_21_lms_array_for_measurement_and_sex(
-            measurement_method=measurement_method, sex=sex, age=age)
-    else:
-        raise ValueError("Incorrect reference supplied")
-    return lms_value_array_for_measurement
-
-
 def generate_centile(z: float, centile: float, measurement_method: str, sex: str, lms_array_for_measurement: list, reference: str) -> list:
     """
     Generates a centile curve for a given reference. 
@@ -339,267 +312,6 @@ def generate_centile(z: float, centile: float, measurement_method: str, sex: str
         # age += (7/365.25) # weekly intervals
     return centile_measurements
 
-def create_uk_who_chart(centile_selection: str=COLE_TWO_THIRDS_SDS_NINE_CENTILES):
-
-    ## user selects which centile collection they want
-    ## If the Cole method is selected, conversion between centile and SDS
-    ## is different as SDS is rounded to the nearest 2/3
-    ## Cole method selection is stored in the cole_method flag.
-    ## If no parameter is passed, default is the Cole method
-
-    centile_collection = []
-
-    if centile_selection == COLE_TWO_THIRDS_SDS_NINE_CENTILES:
-        centile_collection = COLE_TWO_THIRDS_SDS_NINE_CENTILE_COLLECTION
-        cole_method = True
-    else:
-        centile_collection = THREE_PERCENT_CENTILE_COLLECTION
-        cole_method = False
-    
-    ##
-    # iterate through the 4 references that make up UK-WHO
-    # There will be a list for each one
-    ##
-    
-    reference_data = [] # all data for a given reference are stored here: this is returned to the user
-
-    for reference_index, reference in enumerate(UK_WHO_REFERENCES):
-        sex_list: dict = {} # all the data for a given sex are stored here
-        ## For each reference we have 2 sexes
-        for sex_index, sex in enumerate(SEXES):
-            ##For each sex we have 4 measurement_methods
-
-            measurements: dict = {} # all the data for a given measurement_method are stored here
-
-            for measurement_index, measurement_method in enumerate(MEASUREMENT_METHODS):
-                ## for every measurement method we have as many centiles
-                ## as have been requested
-
-                centiles=[] # all generated centiles for a selected centile collection are stored here
-
-                for centile_index, centile in enumerate(centile_collection):
-                    ## we must create a z for each requested centile
-                    ## if the Cole 9 centiles were selected, these are rounded,
-                    ## so conversion to SDS is different
-                    ## Otherwise standard conversation of centile to z is used
-                    if cole_method:
-                        z = rounded_sds_for_centile(centile)
-                    else:
-                        z = sds_for_centile(centile)
-                    
-                    ## Collect the LMS values from the correct reference
-                    lms_array_for_measurement=select_reference_data_for_uk_who_chart(uk_who_reference=reference, measurement_method=measurement_method, sex=sex)
-                    
-                    ## Generate a centile. there will be nine of these if Cole method selected.
-                    ## Some data does not exist at all ages, so any error reflects missing data.
-                    ## If this happens, an empty list is returned.
-                    try:
-                        centile_data = generate_centile(z=z, centile=centile, measurement_method=measurement_method, sex=sex, lms_array_for_measurement=lms_array_for_measurement, reference="uk-who")
-                    except:
-                        print(f"There is no data for {measurement_method} at this age.")
-                        centile_data = []
-
-                    ## Store this centile for a given measurement
-                    centiles.append({"sds": round(z*100)/100, "centile": centile, "data": centile_data})
-                    
-                ## this is the end of the centile_collection for loop
-                ## All the centiles for this measurement, sex and reference are added to the measurements list
-                measurements.update({measurement_method: centiles})
-            
-            ## this is the end of the measurement_methods loop
-            ## All data for all measurement_methods for this sex are added to the sex_list list
-
-            sex_list.update({sex: measurements})
-                
-        ## all data can now be tagged by reference_name and added to reference_data
-        reference_data.append({reference: sex_list})
-        
-    ## returns a list of 4 references, each containing 2 lists for each sex, 
-    ## each sex in turn containing 4 datasets for each measurement_method
-    return reference_data
-
-    """
-
-    structure:
-
-    UK_WHO generates 4 json objects, each structure as below
-
-    uk90_preterm: {
-        male: {
-            height: [
-                {
-                    sds: -2.667,
-                    centile: 0.4
-                    data: [{l: , x: , y: }, ....]
-                }
-            ],
-            weight: [...]
-        },
-        female {...}
-    }
-
-    uk_who_infant: {...}
-    uk_who_child:{...}
-    uk90_child: {...}
-
-    and for the specialist references:
-
-    trisomy_21: {
-        male: {
-            height: [
-                {
-                    sds: -2.667,
-                    centile: 0.4
-                    data: [{l: , x: , y: }, ....]
-                }
-            ],
-            weight: [...]
-        },
-        female {...}
-    }
-
-    """
-def create_trisomy_21_chart(centile_selection: str):
-   ## user selects which centile collection they want
-    ## If the Cole method is selected, conversion between centile and SDS
-    ## is different as SDS is rounded to the nearest 2/3
-    ## Cole method selection is stored in the cole_method flag.
-    ## If no parameter is passed, default is the Cole method
-
-    centile_collection = [] 
-
-    if centile_selection == COLE_TWO_THIRDS_SDS_NINE_CENTILES:
-        centile_collection = COLE_TWO_THIRDS_SDS_NINE_CENTILE_COLLECTION
-        cole_method = True
-    else:
-        centile_collection = THREE_PERCENT_CENTILE_COLLECTION
-        cole_method = False
-    
-    reference_data = {} # all data for a the reference are stored here: this is returned to the user 
-    sex_list: dict = {}
-
-    for sex_index, sex in enumerate(SEXES):
-            ##For each sex we have 4 measurement_methods
-
-        measurements: dict = {} # all the data for a given measurement_method are stored here
-
-        for measurement_index, measurement_method in enumerate(MEASUREMENT_METHODS):
-            ## for every measurement method we have as many centiles
-            ## as have been requested
-
-            centiles=[] # all generated centiles for a selected centile collection are stored here
-
-            for centile_index, centile in enumerate(centile_collection):
-                ## we must create a z for each requested centile
-                ## if the Cole 9 centiles were selected, these are rounded,
-                ## so conversion to SDS is different
-                ## Otherwise standard conversation of centile to z is used
-                if cole_method:
-                    z = rounded_sds_for_centile(centile)
-                else:
-                    z = sds_for_centile(centile)
-                
-                ## Collect the LMS values from the correct reference
-                lms_array_for_measurement=select_reference_data_for_trisomy_21(measurement_method=measurement_method, sex=sex)
-                ## Generate a centile. there will be nine of these if Cole method selected.
-                ## Some data does not exist at all ages, so any error reflects missing data.
-                ## If this happens, an empty list is returned.
-                try:
-                    centile_data = generate_centile(z=z, centile=centile, measurement_method=measurement_method, sex=sex, lms_array_for_measurement=lms_array_for_measurement, reference=TRISOMY_21)
-                except:
-                    print(f"There is no data in {reference} for {measurement_method} at this age.")
-                    centile_data = []
-
-                ## Store this centile for a given measurement
-                centiles.append({"sds": round(z*100)/100, "centile": centile, "data": centile_data})
-                
-            ## this is the end of the centile_collection for loop
-            ## All the centiles for this measurement, sex and reference are added to the measurements list
-            measurements.update({measurement_method: centiles})
-            
-            ## this is the end of the measurement_methods loop
-            ## All data for all measurement_methods for this sex are added to the sex_list list
-
-            sex_list.update({sex: measurements})
-                
-    ## all data can now be tagged by reference_name and added to reference_data
-    reference_data={TRISOMY_21: sex_list}
-    return reference_data
-
-def create_turner_chart(centile_selection: str):
-   ## user selects which centile collection they want
-    ## If the Cole method is selected, conversion between centile and SDS
-    ## is different as SDS is rounded to the nearest 2/3
-    ## Cole method selection is stored in the cole_method flag.
-    ## If no parameter is passed, default is the Cole method
-
-    centile_collection = []
-
-    if centile_selection == COLE_TWO_THIRDS_SDS_NINE_CENTILES:
-        centile_collection = COLE_TWO_THIRDS_SDS_NINE_CENTILE_COLLECTION
-        cole_method = True
-    else:
-        centile_collection = THREE_PERCENT_CENTILE_COLLECTION
-        cole_method = False
-    
-    reference_data = {} # all data for a the reference are stored here: this is returned to the user 
-    sex_list: dict = {}
-
-    for sex_index, sex in enumerate(SEXES):
-            ##For each sex we have 4 measurement_methods
-            ## Turner is female only, but we will generate empty arrays for male
-            ## data to keep all objects the same
-
-        measurements: dict = {} # all the data for a given measurement_method are stored here
-
-        for measurement_index, measurement_method in enumerate(MEASUREMENT_METHODS):
-            ## for every measurement method we have as many centiles
-            ## as have been requested
-
-            centiles=[] # all generated centiles for a selected centile collection are stored here
-
-            for centile_index, centile in enumerate(centile_collection):
-                ## we must create a z for each requested centile
-                ## if the Cole 9 centiles were selected, these are rounded,
-                ## so conversion to SDS is different
-                ## Otherwise standard conversation of centile to z is used
-                if cole_method:
-                    z = rounded_sds_for_centile(centile)
-                else:
-                    z = sds_for_centile(centile)
-                
-                ## Collect the LMS values from the correct reference
-                try:
-                    lms_array_for_measurement=select_reference_data_for_turners(measurement_method=measurement_method, sex=sex)
-                except LookupError:
-                    # there is no data in the reference
-                    lms_array_for_measurement=[]
-                
-                ## Generate a centile. there will be nine of these if Cole method selected.
-                ## Some data does not exist at all ages, so any error reflects missing data.
-                ## If this happens, an empty list is returned.
-                try:
-                    centile_data = generate_centile(z=z, centile=centile, measurement_method=measurement_method, sex=sex, lms_array_for_measurement=lms_array_for_measurement, reference=TRISOMY_21)
-                except:
-                    print(f"There is no data for {measurement_method} at this age.")
-                    centile_data = []
-
-                ## Store this centile for a given measurement
-                centiles.append({"sds": round(z*100)/100, "centile": centile, "data": centile_data})
-                
-            ## this is the end of the centile_collection for loop
-            ## All the centiles for this measurement, sex and reference are added to the measurements list
-            measurements.update({measurement_method: centiles})
-            
-            ## this is the end of the measurement_methods loop
-            ## All data for all measurement_methods for this sex are added to the sex_list list
-
-            sex_list.update({sex: measurements})
-                
-    ## all data can now be tagged by reference_name and added to reference_data
-    reference_data={TURNERS: sex_list}
-    return reference_data
-
 
 
 def rounded_sds_for_centile(centile:float)->float:
@@ -620,29 +332,27 @@ def sds_for_centile(centile: float)->float:
     sds = stats.norm.ppf(centile/100)
     return sds
 
+def lms_value_array_for_measurement_for_reference(
+    reference: str,
+    age: float,
+    measurement_method: str,
+    sex: str,
+    born_preterm: bool
+) -> list:
+    """
+    This is a private function which returns the LMS array for measurement_method and sex and reference
+    It accepts the reference ('uk-who', 'turners-syndrome' or 'trisomy-21')
+    """
 
-def sds_value_for_centile_value(centile: float):
-    ## to be deprecated
-
-    if centile == 0.4:
-        return -2.0 - (2 / 3)
-    elif centile == 2:
-        return -2.0
-    elif centile == 9:
-        return -1 - (1 / 3)
-    elif centile == 25:
-        return 0 - (2 / 3)
-    elif centile == 50:
-        return 0
-    elif centile == 75:
-        return 2 / 3
-    elif centile == 91:
-        return 1 + (1 / 3)
-    elif centile == 98:
-        return 2.0
-    elif centile == 99.6:
-        return 2 + (2 / 3)
+    if reference == UK_WHO:
+        lms_value_array_for_measurement = uk_who_lms_array_for_measurement_and_sex(
+            age=age, measurement_method=measurement_method, sex=sex, born_preterm=born_preterm)
+    elif reference == TURNERS:
+        lms_value_array_for_measurement = turner_lms_array_for_measurement_and_sex(
+            measurement_method=measurement_method, sex=sex, age=age)
+    elif reference == TRISOMY_21:
+        lms_value_array_for_measurement = trisomy_21_lms_array_for_measurement_and_sex(
+            measurement_method=measurement_method, sex=sex, age=age)
     else:
-        # error
-        raise LookupError("SDS could not be calculated from Centile supplied")
-
+        raise ValueError("Incorrect reference supplied")
+    return lms_value_array_for_measurement
